Refactoring to lib.rs
